+++ src/lang_encoders/lang_external_encoder.ml
@@ -26,8 +26,8 @@ open Lang_encoders
 let make params =
   let defaults =
     { External_encoder_format.
-        channels = 2 ;
-        samplerate = 44100 ;
+        channels = Lazy.force Frame.audio_channels ;
+        samplerate = Lazy.force Frame.audio_rate ;
         video = false ;
         header  = true ;
         restart_on_crash = false ;